+++ CHANGELOG.md
@@ -2,6 +2,13 @@ Release Notes
 =============
 
 
+## Next Version
+
+**Breaking Changes**
+
+- `DatabaseIntRepresentable`, `DatabaseInt32Representable`, `DatabaseInt64Representable` and `DatabaseStringRepresentable` have been replaced by the single protocol `DatabaseRawRepresentable` ([documentation](https://github.com/groue/GRDB.swift#swift-enums))
+
+
 ## 0.45.1
 
 Released February 11, 2016